% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{jdfinited}
\alias{jdfinited}
\title{Check if DataFrames and DataArrays packages have been loaded in julia}
\usage{
jdfinited()
}
\value{
single logical
}
\description{
These packages are loaded any time r2j is used to pass data.frames or
any objects containing NAs.
}
\seealso{
Other rjulia: \code{\link{j2r}}, \code{\link{jDo}},
  \code{\link{jloaddf}},
  \code{\link{julia_BigintToDouble}},
  \code{\link{julia_init}}, \code{\link{r2j}},
  \code{\link{rjulia package}}
}
