% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{jDo}
\alias{jDo}
\title{Execute julia code}
\usage{
jDo(cmdstr)
}
\arguments{
\item{cmdstr}{single character}
}
\value{
nothing
}
\description{
Execute julia code
}
\seealso{
Other rjulia: \code{\link{j2r}}, \code{\link{jdfinited}},
  \code{\link{jloaddf}},
  \code{\link{julia_BigintToDouble}},
  \code{\link{julia_init}}, \code{\link{r2j}},
  \code{\link{rjulia package}}
}
