% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{jloaddf}
\alias{jloaddf}
\title{Load the DataFrames and DataArrays packages in juila}
\usage{
jloaddf()
}
\value{
single logical for success or failure
}
\description{
These packages are loaded any time r2j is used to pass data.frames or
any objects containing NAs, but you can load them early if you want.
}
