% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{julia_BigintToDouble}
\alias{julia_BigintToDouble}
\title{Set option for converting julia BigInt values to doubles}
\usage{
julia_BigintToDouble(mode = FALSE)
}
\arguments{
\item{mode}{single logical}
}
\description{
Set option for converting julia BigInt values to doubles
}
