% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{julia_init}
\alias{julia_init}
\title{Initialise Julia}
\usage{
julia_init(disablegc = FALSE)
}
\arguments{
\item{disablegc}{single logical, turn of garbage collection?}
}
\value{
TRUE, invisibly
}
\description{
Initialise Julia
}
