% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{r2j}
\alias{r2j}
\title{Send data to juila}
\usage{
r2j(x, y)
}
\arguments{
\item{x}{R variable: vector, array, factor, list, data.frame}

\item{y}{single character, symbol to use in julia}
}
\value{
nothing
}
\description{
R data types will be translated to the appropriate julia types. For now,
names will be lost.
}
