% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\docType{package}
\name{rjulia package}
\alias{rjulia package}
\alias{rjulia package-package}
\title{Integrating R and Julia -- Calling Julia from R}
\description{
Provides a mechanism to send data to create a running julia proess,
send and receive data to/from that process and run arbitrary code in
the process.
}
