% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{j2r}
\alias{j2r}
\title{Evaluate julia code and return result to R}
\usage{
j2r(expression)
}
\arguments{
\item{expression}{single character}
}
\value{
single character
}
\description{
Evaluate julia code and return result to R
}
