% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{julia_init}
\alias{julia_init}
\title{Start Julia from R,  properly initializing the  R--Julia Interface}
\usage{
julia_init(disablegc = FALSE)
}
\arguments{
\item{disablegc}{single logical, turn of garbage collection?}
}
\value{
TRUE, invisibly
}
\description{
Start Julia from R,  properly initializing the  R--Julia Interface
}
\seealso{
Other rjulia: \code{\link{j2r}}, \code{\link{jDo}},
  \code{\link{jdfinited}}, \code{\link{jloaddf}},
  \code{\link{julia_BigintToDouble}}, \code{\link{r2j}},
  \code{\link{rjulia package}}
}
