% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjulia.R
\name{r2j}
\alias{r2j}
\title{Send data to juila}
\usage{
r2j(x, y)
}
\arguments{
\item{x}{R variable: vector, array, factor, list, data.frame}

\item{y}{single character, symbol to use in julia}
}
\value{
nothing
}
\description{
R data types will be translated to the appropriate julia types. For now,
names will be lost.
}
\examples{
r2j(pi,  "Rpi")
j2r("pi") # Julia's builtin pi
## is there a difference
j2r("Rpi - pi") # typically 0

## integer vector:
r2j(1:10,  "i10")
j2r("i10") ; stopifnot(identical(1:10,  j2r("i10")))

## matrix,  passed to Julia and got back
m <- matrix(exp(-5:6),  3, 4)
r2j(m,  "m")
stopifnot(identical(m,  j2r("m")))
}
\seealso{
Other rjulia: \code{\link{j2r}}, \code{\link{jDo}},
  \code{\link{jdfinited}}, \code{\link{jloaddf}},
  \code{\link{julia_BigintToDouble}},
  \code{\link{julia_init}}, \code{\link{rjulia package}}
}
