\name{rjulia-package}
\alias{rjulia-package}
\alias{rjulia}
\docType{package}
\title{
rjulia is a package that enable R commuication with julia.
}
\description{
rjulia provides an interface between R and Julia,which can convert R and julia object mutually,and run julia script from R.
R is a sophisticated feature-rich statistical software but a bit slow,julia is new advanced computing language,which is fast and llvm-based JIT.so combine R and julia will allow us do better statistical computing.
}
\details{
\tabular{ll}{
Package: \tab rjulia\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-02\cr
License: \tab GPL 2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Yu Gong <armgong@yahoo.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
